requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
comtypes==1.1.11
decorator==4.4.2
ffmpeg==1.4
idna==3.3
imageio==2.18.0
imageio-ffmpeg==0.4.7
moviepy==1.0.3
numpy==1.22.3
Pillow==9.1.0
playsound==1.3.0
proglog==0.1.9
pypiwin32==223
python-dotenv==0.20.0
pyttsx3==2.90
pywin32==303
requests==2.27.1
SpeechRecognition==3.8.1
tqdm==4.64.0
urllib3==1.26.9
